feat(module): add `theme.prefix` option
