+++ docs/content/docs/1.getting-started/2.installation/1.nuxt.md
@@ -363,6 +363,45 @@ export default defineNuxtConfig({
 })
 ```
 
+### `theme.prefix` :badge{label="Soon"}
+
+Use the `theme.prefix` option to configure a prefix for all Tailwind CSS utility classes to avoid naming conflicts with other CSS frameworks or libraries.
+
+First, enable the prefix in your CSS import:
+
+```css [app/assets/css/main.css]
+@import "tailwindcss" prefix(tw);
+@import "@nuxt/ui";
+```
+
+Then configure the same prefix in your Nuxt config:
+
+```ts [nuxt.config.ts]
+export default defineNuxtConfig({
+  modules: ['@nuxt/ui'],
+  css: ['~/assets/css/main.css'],
+  ui: {
+    theme: {
+      prefix: 'tw'
+    }
+  }
+})
+```
+
+This will automatically prefix all Tailwind utility classes and CSS variables in Nuxt UI component themes:
+
+```html
+<!-- Without prefix -->
+<button class="px-2 py-1 text-xs hover:bg-primary/75">Button</button>
+
+<!-- With prefix: tw -->
+<button class="tw:px-2 tw:py-1 tw:text-xs tw:hover:bg-primary/75">Button</button>
+```
+
+::note{to="https://tailwindcss.com/docs/styling-with-utility-classes#using-the-prefix-option" target="_blank"}
+Learn more about using a prefix in the Tailwind CSS documentation.
+::
+
 ### `mdc`
 
 Use the `mdc` option to force the import of Nuxt UI `<Prose>` components even if `@nuxtjs/mdc` or `@nuxt/content` is not installed.
