+++ docs/content/docs/1.getting-started/2.installation/2.vue.md
@@ -658,6 +658,52 @@ export default defineConfig({
 })
 ```
 
+### `theme.prefix` :badge{label="Soon"}
+
+Use the `theme.prefix` option to configure a prefix for all Tailwind CSS utility classes to avoid naming conflicts with other CSS frameworks or libraries.
+
+- Default: `undefined`{lang="ts-type"}
+
+First, enable the prefix in your CSS import:
+
+```css [assets/main.css]
+@import "tailwindcss" prefix(tw);
+@import "@nuxt/ui";
+```
+
+Then configure the same prefix in your Vite config:
+
+```ts [vite.config.ts]
+import { defineConfig } from 'vite'
+import vue from '@vitejs/plugin-vue'
+import ui from '@nuxt/ui/vite'
+
+export default defineConfig({
+  plugins: [
+    vue(),
+    ui({
+      theme: {
+        prefix: 'tw'
+      }
+    })
+  ]
+})
+```
+
+This will automatically prefix all Tailwind utility classes and CSS variables in Nuxt UI component themes:
+
+```html
+<!-- Without prefix -->
+<button class="px-2 py-1 text-xs hover:bg-primary/75">Button</button>
+
+<!-- With prefix: tw -->
+<button class="tw:px-2 tw:py-1 tw:text-xs tw:hover:bg-primary/75">Button</button>
+```
+
+::note{to="https://tailwindcss.com/docs/styling-with-utility-classes#using-the-prefix-option" target="_blank"}
+Learn more about using a prefix in the Tailwind CSS documentation.
+::
+
 ### `inertia`
 
 Use the `inertia` option to enable compatibility with [Inertia.js](https://inertiajs.com/).
