+++ src/runtime/index.css
@@ -6,10 +6,6 @@
 @variant dark (&:where(.dark, .dark *));
 
 @layer base {
-  body {
-    @apply antialiased text-default bg-default scheme-light dark:scheme-dark;
-  }
-
   :root {
     --ui-header-height: --spacing(16);
   }
